BUGFIX: Typecast PageTS Crawler Configuration force_ssl to int (#1095)
Resolves #1000

+++ Classes/Controller/CrawlerController.php
@@ -236,7 +236,7 @@ class CrawlerController implements LoggerAwareInterface
                 $pageId,
                 $urlQuery,
                 $vv['subCfg']['baseUrl'] ?? null,
-                $vv['subCfg']['force_ssl'] ?? 0
+                (int) ($vv['subCfg']['force_ssl'] ?? 0)
             );
 
             if (!$url instanceof UriInterface) {
